// app/interfaces/ILookup.ts
export interface ILookup{
  Suffix: string
  Funder: string
  MembershipNo: string
}

export interface IMemberDetail{
  NumberOfBeneficiaryRecord: string
  MembershipNumber: string
  DependantCode: string
  Funder: string
  Type: string
  Status: string
  BiometricEnrollmentStatus: string
  Surname: string
  FirstName: string
  Initials: string
  Title: string
  Gender: string
  DOB: string
  NationalIDNumber: string
  MaritalStatus: string
  TerminationDate: string
  City: string
  Code: string
  Country: string
  Mobile: string
}